+++ AoE2.netDesktopTests/Form/FormMainPrivate.cs
@@ -2,6 +2,7 @@
 using System.Windows.Forms;
 using AoE2NetDesktop.Tests;
 using System.Drawing;
+using System.ComponentModel;
 
 namespace AoE2NetDesktop.Form.Tests
 {
@@ -28,21 +29,43 @@ namespace AoE2NetDesktop.Form.Tests
                 TestUtilityExt.SetSettings(this, "SelectedIdType", IdType.Profile);
             }
 
-            public async void FormMainOnKeyDown(Keys keys)
+            public async void FormMain_KeyDown(Keys keys)
             {
                 this.Invoke("FormMain_KeyDown", this, new KeyEventArgs(keys));
                 await Awaiter.WaitAsync("FormMain_KeyDown");
             }
 
-            public void FormMainOnMouseDown(MouseEventArgs e)
+            public void FormMain_MouseDown(MouseEventArgs e)
             {
                 this.Invoke("FormMain_MouseDown", this, e);
             }
 
-            public void FormMainOnMouseMove(MouseEventArgs e)
+            public void FormMain_MouseMove(MouseEventArgs e)
             {
                 this.Invoke("FormMain_MouseMove", this, e);
             }
+
+            public void Controls_MouseDown(MouseEventArgs e)
+            {
+                this.Invoke("Controls_MouseDown", this, e);
+            }
+
+
+            public void Controls_MouseMove(MouseEventArgs e)
+            {
+                this.Invoke("Controls_MouseMove", this, e);
+            }
+
+            public void FormMain_MouseClick(MouseEventArgs e)
+            {
+                this.Invoke("FormMain_MouseClick", this, e);
+            }
+
+            public void OnChangeProperty(object sender, PropertyChangedEventArgs e)
+            {
+                this.Invoke("OnChangeProperty", sender, e);
+            }
+            
         }
     }
 }
