+++ AoE2.netDesktopTests/Form/FormSettingsPrivate.cs
@@ -13,6 +13,7 @@ namespace AoE2NetDesktop.Form.Tests
             public Button buttonSetId;
             public CheckBox checkBoxAlwaysOnTop;
             public CheckBox checkBoxHideTitle;
+            public CheckBox checkBoxDrawQuality;
             public Label labelAoE2NetStatus;
             public Label labelErrText;
             public Label labelSettingsName;
@@ -36,6 +37,7 @@ namespace AoE2NetDesktop.Form.Tests
                 buttonSetId = this.GetControl<Button>("buttonSetId");
                 checkBoxAlwaysOnTop = this.GetControl<CheckBox>("checkBoxAlwaysOnTop");
                 checkBoxHideTitle = this.GetControl<CheckBox>("checkBoxHideTitle");
+                checkBoxDrawQuality = this.GetControl<CheckBox>("checkBoxDrawQuality");
                 labelAoE2NetStatus = this.GetControl<Label>("labelAoE2NetStatus");
                 labelErrText = this.GetControl<Label>("labelErrText");
                 labelSettingsName = this.GetControl<Label>("labelSettingsName");
@@ -52,24 +54,46 @@ namespace AoE2NetDesktop.Form.Tests
                 TestUtilityExt.SetSettings(this, "SelectedIdType", IdType.Profile);
             }
 
-            public void FormMainOnMouseDown(MouseEventArgs e)
-            {
-                this.Invoke("FormMain_MouseDown", this, e);
-            }
-
-            public void FormMainOnMouseMove(MouseEventArgs e)
-            {
-                this.Invoke("FormMain_MouseMove", this, e);
-            }
-
-            public void PictureBoxChromaKeyOnClick(EventArgs e)
-            {
-                this.Invoke("PictureBoxChromaKey_Click", this, e);
-            }
+            ///////////////////////////////////////////////////////////////////////
+            // private method
+            ///////////////////////////////////////////////////////////////////////
 
             public void SetChromaKey(string htmlColor)
             {
                 this.Invoke("SetChromaKey", htmlColor);
+            }
+
+            public void ReloadProfileAsync(IdType idtype, string idText)
+            {
+                this.Invoke("ReloadProfileAsync", idtype, idText);
+            }
+
+            public void OnErrorHandler(Exception ex)
+            {
+                this.Invoke("OnErrorHandler", ex);
+            }
+
+            ///////////////////////////////////////////////////////////////////////
+            // Event handlers
+            ///////////////////////////////////////////////////////////////////////
+            public void FormMain_MouseDown(MouseEventArgs e)
+            {
+                this.Invoke("FormMain_MouseDown", this, e);
+            }
+
+            public void FormMain_MouseMove(MouseEventArgs e)
+            {
+                this.Invoke("FormMain_MouseMove", this, e);
+            }
+
+            public void PictureBoxChromaKey_Click(EventArgs e)
+            {
+                this.Invoke("PictureBoxChromaKey_Click", pictureBoxChromaKey, e);
+            }
+
+            public void TextBoxChromaKey_Leave(EventArgs e)
+            {
+                this.Invoke("TextBoxChromaKey_Leave", textBoxChromaKey, e);
             }
         }
     }
