+++ AoE2.netDesktopTests/Form/FormSettingsTests.cs
@@ -1,7 +1,12 @@
 ﻿using AoE2NetDesktop.Tests;
+
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+
 using System;
+using System.Collections.Generic;
 using System.Drawing;
+using System.Net.Http;
+using System.Threading.Tasks;
 
 namespace AoE2NetDesktop.Form.Tests
 {
@@ -26,7 +31,7 @@ namespace AoE2NetDesktop.Form.Tests
             testClass.Shown += async (sender, e) =>
             {
                 await testClass.Awaiter.WaitAsync("FormSettings_Load");
-                testClass.PictureBoxChromaKeyOnClick(e);
+                testClass.PictureBoxChromaKey_Click(e);
 
                 // CleanUp
                 testClass.Close();
@@ -39,6 +44,60 @@ namespace AoE2NetDesktop.Form.Tests
             Assert.AreEqual(expVal, ColorTranslator.FromHtml(testClass.Controler.PropertySetting.ChromaKey));
             Assert.AreEqual(expVal, testClass.pictureBoxChromaKey.BackColor);
             Assert.AreEqual(expVal, ColorTranslator.FromHtml(testClass.textBoxChromaKey.Text));
+            Assert.IsTrue(done);
+        }
+
+        [TestMethod()]
+        public void FormSettingsTestCheckBoxDrawQuality_CheckedChanged()
+        {
+            // Arrange
+            var done = false;
+            var testClass = new FormSettingsPrivate();
+
+            // Act
+            testClass.Shown += async (sender, e) =>
+            {
+                await testClass.Awaiter.WaitAsync("FormSettings_Load");
+                testClass.checkBoxDrawQuality.Checked = true;
+
+                // CleanUp
+                testClass.Close();
+                done = true;
+            };
+
+            testClass.ShowDialog();
+
+            // Assert
+            Assert.IsTrue(testClass.Controler.PropertySetting.DrawHighQuality);
+            Assert.IsTrue(TestUtilityExt.GetSettings<bool>(testClass, "DrawHighQuality"));
+            Assert.IsTrue(done);
+        }
+
+        [TestMethod()]
+        [DataRow("123456", "#123456")]
+        [DataRow("#123456", "#123456")]
+        public void FormSettingsTestTextBoxChromaKey_Leave(string keyValue, string expValTextBox)
+        {
+            // Arrange
+            var done = false;
+            var testClass = new FormSettingsPrivate();
+
+            // Act
+            testClass.Shown += async (sender, e) =>
+            {
+                await testClass.Awaiter.WaitAsync("FormSettings_Load");
+                testClass.textBoxChromaKey.Text = keyValue;
+                testClass.TextBoxChromaKey_Leave(new EventArgs());
+
+                // CleanUp
+                testClass.Close();
+                done = true;
+            };
+
+            testClass.ShowDialog();
+
+            // Assert
+            Assert.AreEqual(expValTextBox, testClass.textBoxChromaKey.Text);
             Assert.IsTrue(done);
         }
 
@@ -295,11 +354,15 @@ namespace AoE2NetDesktop.Form.Tests
         }
 
         [TestMethod()]
-        public void FormSettingsTestButtonSetId_ClickAsyncProfileId()
+        [DataRow(IdType.Steam)]
+        [DataRow(IdType.Profile)]
+        public void FormSettingsTestButtonSetId_ClickAsync(IdType idType)
         {
             // Arrange
             var testClass = new FormSettingsPrivate();
             var done = false;
+
+            TestUtilityExt.SetSettings(testClass, "SelectedIdType", idType);
 
             // Act
             testClass.Shown += async (sender, e) =>
@@ -311,6 +374,34 @@ namespace AoE2NetDesktop.Form.Tests
                 // Assert
                 Assert.AreEqual("   Name: Player1", testClass.labelSettingsName.Text);
                 Assert.AreEqual("Country: Japan", testClass.labelSettingsCountry.Text);
+
+                // CleanUp
+                testClass.Close();
+                done = true;
+            };
+
+            testClass.ShowDialog();
+            Assert.IsTrue(done);
+        }
+
+        [TestMethod()]
+        public void FormSettingsTestButtonSetId_ClickAsyncCatchException()
+        {
+            // Arrange
+            var testClass = new FormSettingsPrivate();
+            var done = false;
+
+
+            // Act
+            testClass.Shown += async (sender, e) =>
+            {
+                await testClass.Awaiter.WaitAsync("FormSettings_Load");
+                testClass.httpClient.ForceTaskCanceledException = true;
+                testClass.buttonSetId.PerformClick();
+                await testClass.Awaiter.WaitAsync("ButtonSetId_ClickAsync");
+
+                // Assert
+                Assert.IsTrue(testClass.labelAoE2NetStatus.Text.Contains("Timeout"));
 
                 // CleanUp
                 testClass.Close();
@@ -378,6 +469,79 @@ namespace AoE2NetDesktop.Form.Tests
             Assert.AreEqual(expValue, testClass.textBoxChromaKey.Text);
             Assert.AreEqual(ColorTranslator.FromHtml(expValue), testClass.pictureBoxChromaKey.BackColor);
             Assert.AreEqual(expValue, TestUtilityExt.GetSettings<string>(testClass, "ChromaKey"));
+        }
+
+        [TestMethod()]
+        [DataRow(IdType.Steam, TestData.AvailableUserSteamId)]
+        [DataRow(IdType.Profile, TestData.AvailableUserProfileIdString)]
+        [DataRow(IdType.Profile, TestData.AvailableUserProfileIdWithoutSteamIdString)]
+        public void ReloadProfileAsyncTest(IdType idtype, string idText)
+        {
+            // Arrange
+            var testClass = new FormSettingsPrivate();
+            var done = false;
+
+            // Act
+            testClass.Shown += async (sender, e) =>
+            {
+                await testClass.Awaiter.WaitAsync("FormSettings_Load");
+                testClass.ReloadProfileAsync(idtype, idText);
+
+                // CleanUp
+                testClass.Close();
+                done = true;
+            };
+
+            testClass.ShowDialog();
+
+            // Assert
+            Assert.IsTrue(done);
+        }
+
+        [TestMethod()]
+        public void ReloadProfileAsyncTestIdTypeNotSelected()
+        {
+            // Arrange
+            var testClass = new FormSettingsPrivate();
+            var done = false;
+
+            // Act
+            testClass.Shown += async (sender, e) =>
+            {
+                await testClass.Awaiter.WaitAsync("FormSettings_Load");
+                testClass.ReloadProfileAsync(IdType.NotSelected, TestData.AvailableUserProfileIdWithoutSteamIdString);
+
+                // CleanUp
+                testClass.Close();
+                done = true;
+            };
+
+            testClass.ShowDialog();
+
+            // Assert
+            Assert.IsTrue(done);
+        }
+
+        private static IEnumerable<object[]> OnErrorHandlerTestData => new List<object[]>
+        {
+            new object[] { new HttpRequestException("404"), "Invalid ID" , Color.Red },
+            new object[] { new HttpRequestException(""), "Server Error", Color.Olive },
+            new object[] { new TaskCanceledException(), "Timeout", Color.Purple },
+        };
+
+        [TestMethod()]
+        [DynamicData(nameof(OnErrorHandlerTestData))]
+        public void OnErrorHandlerTest(Exception ex, string expValueText, Color expValueForeColor)
+        {
+            // Arrange
+            var testClass = new FormSettingsPrivate();
+
+            // Act
+            testClass.OnErrorHandler(ex);
+
+            // Assert
+            Assert.AreEqual(expValueText, testClass.labelAoE2NetStatus.Text);
+            Assert.AreEqual(expValueForeColor, testClass.labelAoE2NetStatus.ForeColor);
         }
     }
 }
