+++ AoE2.netDesktopTests/Form/LabelExTests.cs
@@ -1,0 +1,40 @@
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+using System.Windows.Forms;
+using System.Drawing;
+using System.Collections.Generic;
+using LibAoE2net;
+
+namespace AoE2NetDesktop.Form.Tests
+{
+    [TestClass()]
+    public class LabelExTests
+    {
+        private static IEnumerable<object[]> OnErrorHandlerTestData => new List<object[]>
+        {
+            new object[] { NetStatus.ComTimeout, "Timeout", Color.Purple },
+            new object[] { NetStatus.Connected, "Online", Color.Green },
+            new object[] { NetStatus.Connecting, "Connecting", Color.MediumSeaGreen },
+            new object[] { NetStatus.Disconnected, "Disconnected", Color.Firebrick },
+            new object[] { NetStatus.InvalidRequest, "Invalid ID", Color.Red },
+            new object[] { NetStatus.ServerError, "Server Error", Color.Olive },
+            new object[] { (NetStatus)(-1), string.Empty, new Control().ForeColor },
+        };
+
+        [TestMethod()]
+        [DynamicData(nameof(OnErrorHandlerTestData))]
+
+        public void SetAoE2netStatusTest(NetStatus netStatus, string expText, Color expColor)
+        {
+            // Arrange
+            var testClass = new Label();
+
+            // Act
+            testClass.SetAoE2netStatus(netStatus);
+
+            // Assert
+            Assert.AreEqual(expText, testClass.Text);
+            Assert.AreEqual(expColor, testClass.ForeColor);
+        }
+    }
+}
