+++ AoE2.netDesktopTests/PlotEx/PlayerCountryPlotTests.cs
@@ -1,5 +1,4 @@
-﻿using AoE2NetDesktop.Form;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 using System;
 using System.Collections.Generic;
