+++ AoE2.netDesktopTests/PlotEx/PlotHighlightTests.cs
