+++ AoE2.netDesktopTests/PlotEx/StackedBarGraphDataTests.cs
