+++ AoE2.netDesktopTests/Utility/ColorDialogExTests.cs
@@ -1,11 +1,4 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using AoE2NetDesktop.Form;
-
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Drawing;
 
 namespace AoE2NetDesktop.Form.Tests
